# components/db-manager/pkg/api/docs/swagger.yaml
definitions:
  models.Comment:
    properties:
      audioId:
        description: nullable media id
        type: string
      createdAt:
        type: string
      id:
        type: string
      ownerId:
        type: string
      replies:
        description: comment ids
        items:
          type: string
        type: array
      text:
        type: string
      updatedAt:
        type: string
    type: object
  models.CommentInputAPI:
    properties:
      audioId:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      replies:
        description: UUID strings
        items:
          type: string
        type: array
      text:
        example: Great picture!
        type: string
    type: object
  models.ErrorResp:
    properties:
      error:
        type: string
    type: object
  models.ListResp-models_Comment:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      total:
        type: integer
    type: object
  models.ListResp-models_Location:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Location'
        type: array
      total:
        type: integer
    type: object
...
  models.Location:
    properties:
      address:
        type: string
      createdAt:
        type: string
      descriptionId:
        description: -> comment id
        type: string
      id:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      ownerId:
        type: string
      updatedAt:
        type: string
    type: object
  models.LocationInputAPI:
    properties:
      address:
        example: 221B Baker St, London
        type: string
      descriptionId:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
      lat:
        example: 51.523767
        type: number
      lng:
        example: -0.158555
        type: number
      name:
        example: Grandma’s house
        type: string
    type: object
...
info:
  contact:
    email: guillermo.gc1994@gmail.com
    name: Guillermo Gomez
    url: https://github.com/guigomcha/famquest/
  description: Handles the connection to the DBs in DB Manager. PostgreSQL and MINIO
  license:
    name: Guillermo Gomez GPL V3
  termsOfService: http://swagger.io/terms/
  title: FamQuest DB Manager API
  version: 0.4.0
paths:
  /comments:
    get:
      description: Return paginated comment list
      parameters:
      - description: page size
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListResp-models_Comment'
      summary: List comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Creates a new comment
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CommentInputAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Create comment
      tags:
      - comments
  /comments/{id}:
    delete:
      description: Remove a comment
      parameters:
      - description: comment id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Delete comment
      tags:
      - comments
    get:
      description: Return a single comment
      parameters:
      - description: comment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Get comment by id
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update an existing comment
      parameters:
      - description: comment id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CommentInputAPI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Update comment
      tags:
      - comments
...
  /locations:
    get:
      description: Return paginated location list
      parameters:
      - description: page size
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListResp-models_Location'
      summary: List locations
      tags:
      - locations
    post:
      consumes:
      - application/json
      description: Creates a new location
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LocationInputAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Create location
      tags:
      - locations
  /locations/{id}:
    delete:
      description: Remove a location
      parameters:
      - description: location id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Delete location
      tags:
      - locations
    get:
      description: Return a single location
      parameters:
      - description: location id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Get location by id
      tags:
      - locations
    put:
      consumes:
      - application/json
      description: Update an existing location
      parameters:
      - description: location id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LocationInputAPI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Location'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResp'
      summary: Update location
      tags:
      - locations
...
swagger: "2.0"